Fix indentation in ref usage example
The `conj` is inside the `if` inside the `alter`, so it shouldn’t be indented at the same level as `alter` on the previous line.

I set the new indentation using Vim’s automatic indenter for Lisp. If I use Sublime Text to reindent, I get four spaces of indentation instead, but I think this indentation makes more sense.